internal/dinosql: add map containing all enums
Currently if I want to test for enum membership, I need to write a
custom function and then remember update it every time I add a new
field to an enum.

By placing all of the enums in a map, it is possible to write a
function in a user file in the same package that tests for membership
and does not require updating (since it can just check the map). This
also avoids the need to change the public API.

Fixes #447.