compiler: temp fix for typecast function parameters
In cases where a FuncCall has a TypeCast node as an
argument, resolveCatalogRefs would erroneously output
two parameters for that argument instead of one.

This fix eliminates the duplicate param and allows queries
with typecast function args to compile, but it loses track
of the argument's connection with the function called. This
means the parameters use the generic ColumnN names even when
the function has named parameters.

A future refactor of resolveCatalogRefs to better address
function calls will be necessary to address the name issue.

(this only edits the experimental package and  does not fix
the old dinosql compiler)

Closes #520