mysql: support to infer type of a duplicated arg
Fix for issue with WHERE clause in mysql (old parser) does not parse a
parameters type correctly in certain circumstances.
Such as:
`SELECT * FROM tablex WHERE (id = sqlc.arg(id) OR sqlc.arg(id) = 0);`
This type of statement is a simple way to conditionally filter by input
arg, so that you do not have to define a GetX and a GetXbyID.
(This is a simplified example)

In sqlc v1.4.0 the second id argument gets parsed and added to the
params struct as `ID_2` but is not given a type.
This causes compile-time errors.

Because of the way parameters are parsed and added to the params struct
this fix only works when you define the parameters as sqlc.args(x)
not with `?` or `:x`.