kotlin: remove runtime dependency
The runtime in https://github.com/kyleconroy/sqlc/pull/368 turned out to
be a bad idea. It's pretty much unnecessary and adds quite a bit to the
complexity of the generated code. It's _possible_ a runtime would be
useful in the future. At that point we can add that back in