Add support for globs in overrides.
This lets us say things like '*.*_ksuid' should be of type
'github.com/segmentio/ksuid.KSUID', instead of having to list every
single column in every single table.

We can even say '*.*.*_ksuid' for every column, in every table, in every
schema.

Now, ideally we would do a type replace on 'binary(20)', however that
gets turned into a char for at least the mysql code, which means that we
never get to see 'binary(20)', and so can't do a replacement on it.

This makes use of the github.com/gobwas/glob package for the globbing,
and it pretty much Just Works the way you expect it to.