Only check for duplicate c.id on query outputs.
That is, we want to deduplicate the same column from showing up multiple
times, and thus we want to avoid adding a suffix, when we are getting
the results from a query, however we most definitely do not want to do
this for query parameters.

Especially since the logic for handling this in the query parameters is
missing the deduplication, but also because the database itself expects
the same number of parameters as we have placeholders in the query.