Make seen use fieldName instead of colName.
fieldName is after any mangilng to make it fit go variable standards,
and since that is the name that we are actually using, that is the name
that we should be checking for conflicts.

This only matters when we have multiple fields that differ only by
things that are changed when we are converting to the go variable
standards.