fix copyfrom methods on interface not matching generated methods
The methods generated for `copyfrom` queries correctly define a slice
of inputs on the function signatures, but the interface code that was
generated did not. Thus, the interface was not usable when any query
used `copyfrom`. This merge request updates the interface code template
to correctly use `.Arg.SlicePair` when building the interface.

Signed-off-by: Alex Dulin <alex@morningconsult.com>