fix bug where datatype is read from
See https://github.com/kyleconroy/sqlc/pull/1460#discussion_r817029197

DataType is never filled in so this comparison is buggy. This change mirrors
what the go gen.go does instead.

Deleting a field from proto can be not safe for compat, so we could use
https://developers.google.com/protocol-buffers/docs/proto3#reserved,
but I think these protos are transient so it should be ok.