Add sqlc.slice() to support IN clauses in MySQL (#695)
This feature (currently MySQL-specific) allows passing in a slice to an
IN clause. Adding the new function sqlc.slice() as opposed to overloading
the parsing of "IN (?)" was chosen to guarantee backwards compatibility.

   SELECT * FROM tab WHERE col IN (sqlc.slice("go_param_name"))

This commit is based on https://github.com/kyleconroy/sqlc/pull/1312 by
Paul Cameron. I just rebased and did some cleanup.

Co-authored-by: Paul Cameron <cameronpm@gmail.com>