internal/cmd: fix an exit code buglet
It seems the previous code intended to interpret the error from
rootCmd.ExecuteContext() as an *exec.ExitError. That's not the way
it worked in practice, though. Instead, the typecheck on
err.(*exec.ExitError) used the error from the call to tracer.Start().
This is due to the scoping rules of if statements.

With this commit, cmd.Do() now attempts to interpret the command
execution error as an *exec.ExitError. It also prints the error
message to stderr, since the caller in ./cmd/sqlc.main doesn't do
that.

See: https://go.dev/ref/spec#Blocks