feat(cmd/sqlc): Add the vet subcommand (#2344)
This is a proposal to add a new subcommand that runs queries through a set of lint rules.
Rules

Rules are defined in the configuration file. They consist of a name, message, and an expression. If the expression evaluates to true, an error is reported. These expressions are evaluated using cel-go.

Each expression has access to a query object, which is defined as the following struct:

---------

Co-authored-by: Kyle Conroy <kyle@sqlc.dev>