feat(sql/ast): Render AST to SQL (#2815)
Add a new Format method to AST nodes to support writing an AST tree back to SQL. The rendered SQL is checked to be equivalent to the input SQL, not exact. The tests current run against all query.sql files in pgx/v5 directories. As support is improved, we'll expand to all PostgreSQL queries.

The hope is to eventually use this to power a sqlc fmt subcommand, but that is a ways off from happening.