Added the ability to add and remove necessary imports at the generation level
using the package "golang.org/x/tools/imports "

This solves the problem in the same way as for packet separation. In the case when, for example, a package of models has been added to the interfaces, but it is not used there.

A problem was also identified due to the fact that the package was not added to the interfaces "github.com/jackc/pgx/v5/pgtype "

But the standard tool fixes these problems on its own