Avoid updating target file mtime in 'generate'
If the target file exists on disk already, compare its contents
to the generated source in memory: only update the target file
on disk if the contents differ or if the target file is missing.

This should allow build tools downstream of sqlc to make conditional
decisions based on whether the generated code files have been updated
(e.g. mockgen).